+++ tests/Fixture/Logger/Psr3LoggerAdapter.php
@@ -1,0 +1,147 @@
+<?php
+
+namespace YoutubeDownloader\Tests\Fixture\Logger;
+
+use Psr\Log\LoggerInterface;
+use YoutubeDownloader\Logger\Logger;
+
+/**
+ * A simple PSR-3 logger as a compatibility proof for the Logger interface
+ */
+class Psr3LoggerAdapter implements LoggerInterface, Logger
+{
+	/**
+	 * @var YoutubeDownloader\Logger\Logger
+	 */
+	private $logger;
+
+	/**
+	 * @param YoutubeDownloader\Logger\Logger $logger
+	 *
+	 * @return void
+	 */
+	public function __construct(Logger $logger)
+	{
+		$this->logger = $logger;
+	}
+
+	/**
+	 * System is unusable.
+	 *
+	 * @param string $message
+	 * @param array $context
+	 * @return null
+	 */
+	public function emergency($message, array $context = array())
+	{
+		return $this->logger->emergency($message, $context);
+	}
+
+	/**
+	 * Action must be taken immediately.
+	 *
+	 * Example: Entire website down, database unavailable, etc. This should
+	 * trigger the SMS alerts and wake you up.
+	 *
+	 * @param string $message
+	 * @param array $context
+	 * @return null
+	 */
+	public function alert($message, array $context = array())
+	{
+		return $this->logger->alert($message, $context);
+	}
+
+	/**
+	 * Critical conditions.
+	 *
+	 * Example: Application component unavailable, unexpected exception.
+	 *
+	 * @param string $message
+	 * @param array $context
+	 * @return null
+	 */
+	public function critical($message, array $context = array())
+	{
+		return $this->logger->critical($message, $context);
+	}
+
+	/**
+	 * Runtime errors that do not require immediate action but should typically
+	 * be logged and monitored.
+	 *
+	 * @param string $message
+	 * @param array $context
+	 * @return null
+	 */
+	public function error($message, array $context = array())
+	{
+		return $this->logger->error($message, $context);
+	}
+
+	/**
+	 * Exceptional occurrences that are not errors.
+	 *
+	 * Example: Use of deprecated APIs, poor use of an API, undesirable things
+	 * that are not necessarily wrong.
+	 *
+	 * @param string $message
+	 * @param array $context
+	 * @return null
+	 */
+	public function warning($message, array $context = array())
+	{
+		return $this->logger->warning($message, $context);
+	}
+
+	/**
+	 * Normal but significant events.
+	 *
+	 * @param string $message
+	 * @param array $context
+	 * @return null
+	 */
+	public function notice($message, array $context = array())
+	{
+		return $this->logger->notice($message, $context);
+	}
+
+	/**
+	 * Interesting events.
+	 *
+	 * Example: User logs in, SQL logs.
+	 *
+	 * @param string $message
+	 * @param array $context
+	 * @return null
+	 */
+	public function info($message, array $context = array())
+	{
+		return $this->logger->info($message, $context);
+	}
+
+	/**
+	 * Detailed debug information.
+	 *
+	 * @param string $message
+	 * @param array $context
+	 * @return null
+	 */
+	public function debug($message, array $context = array())
+	{
+		return $this->logger->debug($message, $context);
+	}
+
+	/**
+	 * Logs with an arbitrary level.
+	 *
+	 * @param mixed $level
+	 * @param string $message
+	 * @param array $context
+	 * @return null
+	 */
+	public function log($level, $message, array $context = array())
+	{
+		return $this->logger->log($leverl, $message, $context);
+	}
+}